+++ src/runtime/io.c
@@ -27,7 +27,7 @@ int read_in_bf(size_t *val){
     return EXIT_SUCCESS;
 }
 
-int read_ws_command_char(FILE *fptr, size_t *out_char){
+int read_ws_command_char(FILE *fptr, int *out_char){
     do {
         *out_char = fgetc(fptr);
         if (*out_char == EOF) {
